style(tests,tools): clear the remaining widened-scope lint findings
The last eight findings clang-tidy-18 reports once tests/ and tools/ are in
scope. All fixed at the cause.

mdl_export_jof.c: jof_one() exceeded the size threshold in its new translation
unit -- it owned the source buffer, two scratch arenas and the output file, with
four cleanup paths between them. Split by resource, the same way
ra8_fmt_atlas_produce was: jof_produce_page() now owns both scratch arenas and
releases them on every exit, leaving jof_one() owning the source buffer and the
file. That also separates two failures the old code had folded into one ternary
-- a zero work_cap and a failed fopen now return their own codes instead of
being told apart after the fact.

test_ra8_ssie_io.c: k_ssie_io_lit_40 was a value-derived name the rename lanes
missed, and its uint32_t width made the uint8_t loop index narrower than its own
bound. Renamed to k_ssie_src_words with the role documented (deliberately larger
than the 32-entry FIFO, so a write is bounded by reported space, not by supplied
data) and moved to a right-sized enum.

test_ra8_usb_pmsc_scsi_cov.c: k_ra8_pmsc_cdb_max_len is a uint16_t production
constant, so the index widens to match rather than the constant narrowing.

test_ra8_jof_probe_dims.c: the JPEG SOI bytes become a byte array beside
s_png_sig and the fourCC tags, so the file states its magics one way.

mdl_extract.c, sim_args.c: the conditional operator promotes both arms to int,
so a cast on one arm does not stop the store narrowing. Cast the whole
expression; both arms were already char-valued.

sim_run.c: a constant's doc comment wrote a path placeholder as <record_dir>,
which Doxygen parsed as an unknown HTML tag and the docs gate reported as a
warning. Written as inline code instead.

+++ tests/test_ra8_usb_pmsc_scsi_cov.c
@@ -240,7 +240,7 @@ static void setup_with(const ra8_usb_pmsc_storage_t* storage)
  */
 static void seed_rw_cdb(uint32_t lba, uint16_t block_count)
 {
-  for (uint8_t i = 0U; i < k_ra8_pmsc_cdb_max_len; ++i) {
+  for (uint16_t i = 0U; i < (uint16_t)k_ra8_pmsc_cdb_max_len; ++i) {
     s_usb_pmsc_state.cbw_cdb[i] = 0U;
   }
   s_usb_pmsc_state.cbw_cdb[k_test_cdb_lba_msb] =
